+++ core/ui/src/commonMain/kotlin/com/whatever/caramel/core/ui/picker/model/DateUiState.kt
@@ -1,9 +1,9 @@
 package com.whatever.caramel.core.ui.picker.model
 
 import com.whatever.caramel.core.util.DateUtil
+import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalDateTime
 
-// @ham2174 FIXME : DateMonthPicker 또는 DatePicker 의 상태 변경시 DateUiState / DateMonthUiState 로 분리
 data class DateUiState(
     val year: Int,
     val month: Int,
@@ -37,3 +37,6 @@ data class DateUiState(
             )
     }
 }
+
+fun DateUiState.toLocalDate(): LocalDate =
+    LocalDate(year, month, day)
